Separate tank out into module
